Add 'Clone' button to the Repository viewer activity
This provides some nice integration with Agit (the Git client app) which
has support for doing full clones of git repos onto the SD card, but no
GitHub integration.

The button launches the org.openintents.git.clone.PREPARE intent which
will give the user a chance to review the details of the checkout (target
dir, bare/not-bare, etc) before starting the clone proper. The button's
hidden if no app can be found to handle the intent (e.g. if Agit isn't
installed).

We have to pass the source-uri for doing the clone, but the GitHub api
doesn't provide the clone url - only the project url, so we work it out
from that, using an ssh url if the repo is private.

I have to apologise but unfortunately I wasn't able to get Gh4a to compile
in the time I had available (I couldn't find the mygson dependency for
https://github.com/slapperwan/github-java-sdk ) so I've been coding blind
to some extent, but I've already tested very similar code against Hubroid,
so I think it should work just fine - please let me know if there are any
problems!