Scale down gravatar images.
Github provides pretty large avatar images, which, when stored unscaled,
can use a lot of memory. Scale them down to the maximum size we use
anywhere (60dp - in the user fragment), if they are larger than that.

Should avoid OOM exceptions, as it brings down the heap size of the
cache significantly (I've seen about 70% here).