Kill the ToolBar for now.
The ScrollView must be at the root of the view hierarchy for the
swipe-to-refresh scroll detection to work (albeit that could also be
done by overriding canChildScrollUp), and more importantly, it doesn't
actually do anything. We ultimately should think whether we should use a
navigation drawer for the (self) user activity.