Fix error view handling.
Fix an exception being thrown for calling supportRequestWindowFeature()
after adding content by no longer requesting FEATURE_NO_TITLE (as this
is done as part of the appcompat theme already), and add a few missing
cases where hasErrorView() wasn't checked correctly.

At some point we should reimplement this to behave like the example
under 'Connectivity' in [1], that is, inflate the view hierarchy and
just hide it under a 'error, try again' kind of view.

[1] http://www.google.com/design/spec/patterns/errors.html#errors-user-input-errors