Support Github's authorization API changes.
Full tokens for existing authorizations aren't returned anymore now, so
we need to create one token per device, using the new fingerprint
property and a more descriptive note. Ideally, we'd also kill the
authorization on logout, but that doesn't seem possible with just the
token.
Also auto-logout on 401 to make sure expired tokens are invalidated
automatically.