Fix handling of ordered lists for API level >= 24.
Html.fromHtml() used to not handle lists at all, so we had our own
handling in a custom tag handler. Since N, it _does_ handle lists, but
only unordered ones, so our special handling broke because we don't get
to see the <li> tags anymore.
As there's no way to get a hook into the framework class, import the
necessary pieces of it and insert our special tag handling that used to
live in the custom TagHandler instance into it.

Closes #415