Add comment text quoting
It works by adding an additional menu item to the actions menu when user selects text in comments. This allows for easy quoting of any part of the comment.

The quote insertion code appends the selected text at the end of the comment input field. It also makes sure that each quoted line is prefixed with quote char to appear correctly when rendered.

Additional code was added to make sure that there is at least one empty line above and under the quote text when adding it to comment input field.