Fix filtering by label.
Keyword is label:, not labels:, and concatenating labels with comma
doesn't actually work. One can add multiple label: keywords, but that
does and AND combination, not an OR one. Since that doesn't match the
multi select semantics, drop the multi selection.
Also allow filtering for 'unlabeled' and 'unset milestone' to keep
consistency.

Closes #451.