Fix "Newest" sort option not getting unhighlighted
If menu item has "checked" set to true it doesn't get unhighlighted when users select other itmes in the drawer. That's because NavigationView has unhighlighting behavior implemented in a way which remembers previous selections but it doesn't check for the initial state.

To fix this issue we have to use the item checking method of the NavigationView instead of the individual items.