Change title of search action in repository activity
Simply calling the action "Search" is a bit too broad since it doesn't specify exactly what the user would be searching for. Is it issues, code, content of readme, files?

This commit fixes the issue by making the action exactly specify what it can be used to search for, which is "Search code".