Add shortcuts to enter labels and milestone management screens
This commits adds neutral buttons to the dialogs displayed in the issue edit activity when selecting labels or milestones. Pressing on these buttons opens corresponding screens and if something is modified by the user they notify the calling activities to reload the data.

To implement this change I had to make a small modification in the `IssueMilestoneEditActivity`. It now uses the `finish()` method after successful actions instead of starting new issue list activity. This is the same behavior as the one used in `IssueEditActivity`. I made the necessary changes to make sure that the parent milestones list reloads its data if it was the one that opened the edit screen.