Fix disappearing comment bottom sheet under some conditions
Before this change the bottom sheet would disappear under these conditions:
- The device was in landscape orientation
- The user was in screen with multiple toolbar tabs

The problem was that whenever we set toolbar to scrollable, then for an unknown reason it would get pushed out of view.

The solution turned out to be pretty simple. We don't want to modify the scrollable state based on locked state anymore because scrolling is now disabled by disabling the whole behavior