Fix line highlight marking incorrect lines
If someone shared a link to a removed line located directly above multiple added lines then the target position would be incorrect. This is caused by the fact that diff position for line of one type was not changing while the other was.

Solution for the problem is to set start and end line positions only once so they aren't pushed down.