Fix search expanded state being restored incorrectly
The expanded state was restored based on whether the search query was set to `null` or not. This method was not working correctly due to one issue: After collapsing, the search view's `onQueryTextChange` method was called again with an empty string as a parameter value, which in turn resulted in search view getting incorrectly expanded after restoring the state.

The solution was to add another field which tracks whether the search view should be expanded separately from the value of the search query.