Directly link to review comments from news feed
In order to implement this improvement I had to add an additional
parameter to UrlLoadTask and ProgressDialogFragment. It is telling them
whether the calling activity should be closed once the dialog is
dismissed or operation finishes.

Without that change the activity with news feed would be automatically
closed after review is loaded, which is not what we want. We want the
user to be able to return here by pressing back key.