Change order of issue properties in edit screen
(And also in issue header)

I've made this change because editing labels is the most commonly
performed action from these in settings screen. (At least from my
experience.) Currently the option to do so is located as a last action
which requires scrolling to reach and is really annoying. Now instead
milestones are positioned as last option because these seem to be the
least commonly used action.