Fix crash when loading avatar for notification
If this handler was accessed from the background thread of notifications job then the application would crash. This could happen if the user didn't start the app for long time (application got removed from memory) and then a notification job attempted to load images using this handler. This in result created it from wrong thread and resulted in crash.