Avoid TransactionTooLargeException with CommitFragment.
It needs to get the Commit object as fragment argument, but that object
can grow extremely large depending on the number of patches attached to
it, which in turn can cause TransactionTooLargeExceptions being thrown
when the arguments are put in the activity's state bundle. Avoid that by
storing the object in compressed form.