Do not do reply-group auto-selection if there are multiple groups
Before this commit auto-selection would kick-in when there were
multiple groups, but only one of them had a reply button. (For other
groups replying was disabled)

Now in that situation we won't do auto-selection and let the user to
manually select the (only) group they want to reply to when they
decide to do so.