Immediately trigger reload after changing search mode.
Otherwise, when changing mode while a search is running, there can be a
mismatch between view holder and item types: when the pending search
(for the old type) finally finishes, the view holders will be generated
for the new type while the data items are of the old type.
Reloading avoids that problem as it cancels off the old pending load if
there was any.