Prevent white splash on cold start
Currently gh4a sets the user theme in BaseActivity::onCreate() but still the white splash is frequently shown on budget devices during a cold start of the app, which is extremely annoying in case the dark theme is used in the app and in the entire UI of the device.

This patch makes it so that a transparent default theme is shown before the actual theme is applied. On a slower device nothing will be shown after the user launched the app until the main screen is presented so one might think that could perplex some users, but I think they are used to their device being slow and encounter such pauses regularly. On a fast device there'll be no perceivable difference to the old behavior.