Fix URL handling activities getting lost sometimes.
If BrowseFilter was called with FLAG_ACTIVITY_NEW_TASK set, it became
the root of a new task, but since it's excluded from recents, the task
got lost once the user moved away from it, e.g. via recents. To avoid
that, proxy the flags BrowseFilter was called with to the resolved
activity, so the started activity gets started in a task independent of
BrowseFilter.

Fixes #946